The axis parameter can also be negative, counting axes backwards from the inner most dimension.

+++ include/ndhist/ndhist.hpp
@@ -53,7 +53,7 @@ class ndhist
     GetBinContentArray();
 
     boost::numpy::ndarray
-    GetEdgesArray(size_t axis=0);
+    GetEdgesArray(int axis=0);
 
   private:
     ndhist() {};
